replace misplaced files
